feat: turn spinviewer to typescript
